app/requirements.txt
discord-webhook
langchain
faiss-cpu
ollama
python-dotenv
requests
openai
aiohttp
asyncio